Tested the database
Because:
- The database needs to be connected properly to the app

I created a button, which adds test data to the database.
I renamed the package name from vpmanagaer to vpmanager.
