+++ .idea/misc.xml
@@ -8,6 +8,7 @@
         <entry key="..\:/Users/stron/AndroidStudioProjects/ase-ws-21-vp-manager/app/src/main/res/layout/activity_find_study.xml" value="0.5541666666666667" />
         <entry key="..\:/Users/stron/AndroidStudioProjects/ase-ws-21-vp-manager/app/src/main/res/layout/activity_home.xml" value="0.5541666666666667" />
         <entry key="..\:/Users/stron/AndroidStudioProjects/ase-ws-21-vp-manager/app/src/main/res/layout/activity_personal_account.xml" value="0.5541666666666667" />
+        <entry key="..\:/Users/user/StudioProjects/ase-ws-21-vp-manager/app/src/main/res/layout/activity_home.xml" value="0.18115942028985507" />
       </map>
     </option>
   </component>
