--- app/src/main/java/com/example/vpmanagaer/createStudyActivity.java
@@ -1,6 +1,0 @@
-package com.example.vpmanagaer;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class createStudyActivity extends AppCompatActivity {
-}
