--- app/src/main/java/com/example/vpmanagaer/findStudyActivity.java
@@ -1,6 +1,0 @@
-package com.example.vpmanagaer;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class findStudyActivity extends AppCompatActivity {
-}
