--- app/src/main/java/com/example/vpmanagaer/personalAccountActivity.java
@@ -1,6 +1,0 @@
-package com.example.vpmanagaer;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class personalAccountActivity extends AppCompatActivity {
-}
