+++ app/src/main/java/com/example/vpmanager/findStudyActivity.java
@@ -1,0 +1,6 @@
+package com.example.vpmanager;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class findStudyActivity extends AppCompatActivity {
+}
