+++ app/src/main/java/com/example/vpmanager/personalAccountActivity.java
@@ -1,0 +1,6 @@
+package com.example.vpmanager;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class personalAccountActivity extends AppCompatActivity {
+}
