+++ app/src/main/res/layout/activity_home.xml
@@ -80,6 +80,13 @@
                     android:layout_alignParentEnd="true"/>
             </RelativeLayout>
 
+            <Button
+                android:id="@+id/dbTestButton"
+                android:layout_width="150dp"
+                android:layout_height="50dp"
+                android:text="TestDB"
+                android:layout_gravity="center"/>
+
         </LinearLayout>
 
     <LinearLayout
